Feature: #34 one-click complete all active todos (#35)
* docs(spec): add TL-0011 specification in documentation

Refs: #34

* feat(todo-service): add completeAllActiveTodos

Refs: #31

* test(todo-service): add unit test for completeAllActiveTodos

Refs: #31

* feat(app): Add 'complete all active todos' functionality

Introduces a feature to mark all active todos as completed, with a new checkbox UI and corresponding styles. Adds logic for determining when todos are fully completed and updates the component state accordingly. Enhances test coverage for the new functionality and improves test selectors for clarity. Includes updates to mock data setup and service methods.

Refs: #31

* feat(add-todo): implement one-click complete active todos

Refs: #34

* test(UT): add unit test for complete active todos

Refs: #34

* test(e2e): add e2e test for TL-0011

Refs: #34

* test(add-todo): fix unit test for add-todo components

Refs: #34

---------

Co-authored-by: idembele70 <idembele70@gmail>